version 20131125 15:30 scutLaoYi ----------------------------------- fix the bug of ajax. Now enter the second level page of Proxy(not the proxy_list, but the one called temp), \ you can search the result by province. Next time try to complete the function of proxy search(by product, by material, by function and so on).
fix the bug of javascript.
Now when you change the selection of product type, the following form will change currently.
Just for the second level page of proxy, not for the third level page of proxy\
(It means that the same bug still exists in third level page of proxy called proxy_submit).