version 20131125 19:00 scutLaoYi
move all javascript for proxy type selection to a outer js file called
"proxySelection.js", including proxy_submit page and proxy_search page.

Now when someone want to call the javascript, he can just added a line
at the top of the view(ctp) file like this:
	<?php echo $this->Html->script('proxySelection'); ?>
and the file will be called and all code in that file will be used.

The parameters for the js function "setOption" changed!
Added a new paremeter called "firstOption", which will be added as the
first option when create some selection.

Next time try to fix the bug of proxy_submit.
When it is used to edit a existed record, the setting of proxy type
can't recover currectly.